refactor: moved job board assets to 2025 data
