+++ src/data/2025/assets/images/job-board/ccs-logo.png
