+++ src/data/2025/assets/images/job-board/compass-logo.png
