+++ src/data/2025/assets/images/job-board/dte-logo.png
