+++ src/data/2025/assets/images/job-board/google-logo.png
