+++ src/data/2025/assets/images/job-board/ibm-logo.png
